Fixed some type conversion warnings from VS

+++ osx/06_diffuse_lighting/source/main.mm
@@ -248,7 +248,7 @@ static void CreateInstances() {
 
     ModelInstance hMid;
     hMid.asset = &gWoodenCrate;
-    hMid.transform = translate(-6,0,0) * scale(2,1,0.8);
+    hMid.transform = translate(-6,0,0) * scale(2,1,0.8f);
     gInstances.push_back(hMid);
 }
 
@@ -340,14 +340,14 @@ static void Update(float secondsElapsed) {
 
 
     //rotate camera based on mouse movement
-    const float mouseSensitivity = 0.1;
+    const float mouseSensitivity = 0.1f;
     int mouseX, mouseY;
     glfwGetMousePos(&mouseX, &mouseY);
     gCamera.offsetOrientation(mouseSensitivity * mouseY, mouseSensitivity * mouseX);
     glfwSetMousePos(0, 0); //reset the mouse, so it doesn't go out of the window
 
     //increase or decrease field of view based on mouse wheel
-    const float zoomSensitivity = -0.2;
+    const float zoomSensitivity = -0.2f;
     float fieldOfView = gCamera.fieldOfView() + zoomSensitivity * (float)glfwGetMouseWheel();
     if(fieldOfView < 5.0f) fieldOfView = 5.0f;
     if(fieldOfView > 130.0f) fieldOfView = 130.0f;
@@ -412,10 +412,10 @@ void AppMain() {
     gLight.intensities = glm::vec3(1,1,1); //white
 
     // run while the window is open
-    double lastTime = glfwGetTime();
+    float lastTime = (float)glfwGetTime();
     while(glfwGetWindowParam(GLFW_OPENED)){
         // update the scene based on the time elapsed since last update
-        double thisTime = glfwGetTime();
+        float thisTime = (float)glfwGetTime();
         Update(thisTime - lastTime);
         lastTime = thisTime;
